+++ test/pubsub_no_strand.cpp
@@ -44,7 +44,7 @@ BOOST_AUTO_TEST_CASE( pub_qos0_sub_qos0 ) {
         (bool sp, std::uint8_t connack_return_code) {
             MQTT_CHK("h_connack");
             BOOST_TEST(sp == false);
-            BOOST_TEST(connack_return_code == MQTT_NS::connect_return_code::accepted);
+            BOOST_TEST(connack_return_code == static_cast<std::uint8_t>(MQTT_NS::connect_return_code::accepted));
             pid_sub = c->subscribe("topic1", MQTT_NS::qos::at_most_once);
             return true;
         });
@@ -85,7 +85,7 @@ BOOST_AUTO_TEST_CASE( pub_qos0_sub_qos0 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_most_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_most_once));
             c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_most_once);
             return true;
         });
@@ -190,7 +190,7 @@ BOOST_AUTO_TEST_CASE( pub_qos1_sub_qos0 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_most_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_most_once));
             pid_pub = c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_least_once);
             return true;
         });
@@ -296,7 +296,7 @@ BOOST_AUTO_TEST_CASE( pub_qos2_sub_qos0 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == 1);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_most_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_most_once));
             pid_pub = c->publish("topic1", "topic1_contents", MQTT_NS::qos::exactly_once);
             return true;
         });
@@ -396,7 +396,7 @@ BOOST_AUTO_TEST_CASE( pub_qos0_sub_qos1 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_least_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_least_once));
             c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_most_once);
             return true;
         });
@@ -509,7 +509,7 @@ BOOST_AUTO_TEST_CASE( pub_qos1_sub_qos1 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_least_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_least_once));
             pid_pub = c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_least_once);
             return true;
         });
@@ -624,7 +624,7 @@ BOOST_AUTO_TEST_CASE( pub_qos2_sub_qos1 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_least_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_least_once));
             pid_pub = c->publish("topic1", "topic1_contents", MQTT_NS::qos::exactly_once);
             return true;
         });
@@ -725,7 +725,7 @@ BOOST_AUTO_TEST_CASE( pub_qos0_sub_qos2 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::exactly_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::exactly_once));
             c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_most_once);
             return true;
         });
@@ -838,7 +838,7 @@ BOOST_AUTO_TEST_CASE( pub_qos1_sub_qos2 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::exactly_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::exactly_once));
             pid_pub = c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_least_once);
             return true;
         });
@@ -953,7 +953,7 @@ BOOST_AUTO_TEST_CASE( pub_qos2_sub_qos2 ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::exactly_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::exactly_once));
             pid_pub = c->publish("topic1", "topic1_contents", MQTT_NS::qos::exactly_once);
             return true;
         });
@@ -1054,7 +1054,7 @@ BOOST_AUTO_TEST_CASE( publish_function ) {
             MQTT_CHK("h_suback");
             BOOST_TEST(packet_id == pid_sub);
             BOOST_TEST(results.size() == 1U);
-            BOOST_TEST(*results[0] == MQTT_NS::qos::at_most_once);
+            BOOST_TEST(*results[0] == static_cast<std::uint8_t>(MQTT_NS::qos::at_most_once));
             c->publish("topic1", "topic1_contents", MQTT_NS::qos::at_most_once);
             return true;
         });
